Move elements to ion-elements
