+++ cmd/avp/main.go
@@ -5,18 +5,18 @@ import (
 	"net/http"
 	_ "net/http/pprof"
 
+	"github.com/pion/ion-elements/elements"
 	conf "github.com/pion/ion/pkg/conf/avp"
 	"github.com/pion/ion/pkg/discovery"
 	"github.com/pion/ion/pkg/log"
 	"github.com/pion/ion/pkg/node/avp"
 	"github.com/pion/ion/pkg/process"
-	"github.com/pion/ion/pkg/process/elements"
 	"github.com/pion/ion/pkg/process/samples"
 	"github.com/pion/ion/pkg/proto"
 )
 
-func getDefaultElements(id string) map[string]elements.Element {
-	de := make(map[string]elements.Element)
+func getDefaultElements(id string) map[string]process.Element {
+	de := make(map[string]process.Element)
 	if conf.Pipeline.WebmSaver.Enabled && conf.Pipeline.WebmSaver.DefaultOn {
 		webm := elements.NewWebmSaver(elements.WebmSaverConfig{
 			ID:   id,
@@ -27,7 +27,7 @@ func getDefaultElements(id string) map[string]elements.Element {
 	return de
 }
 
-func getTogglableElement(msg proto.ElementInfo) (elements.Element, error) {
+func getTogglableElement(msg proto.ElementInfo) (process.Element, error) {
 	switch msg.Type {
 	case elements.TypeWebmSaver:
 		return elements.NewWebmSaver(elements.WebmSaverConfig{
