--- pkg/process/elements/webm_test.go
@@ -1,10 +1,0 @@
-package elements
-
-import (
-	"testing"
-)
-
-func TestWebMSaver(t *testing.T) {
-	saver := NewWebmSaver("id")
-	saver.Close()
-}
